+++ src/main/resources/mybatis/mappers/NotificationMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.woosan.hr_system.notification.dao.NotificationDAO">
+<mapper namespace="notification">
     <resultMap id="NotificationResultMap" type="com.woosan.hr_system.notification.model.Notification">
         <id property="notificationId" column="notification_id"/>
         <result property="employeeId" column="employee_id"/>
